Add files via upload
This is an advanced Excel project that shows an alternative to the stock market app, here you can track your stocks, check all your transactions, and automate your transaction history after buying or selling using an app script from Google or using Macros from Microsoft excel. This was one of the ways I started investing and is a sample of the transactions. The file is not available to be distributed or shared only to view and read codes.